separate location schema
